+++ vg/vg_origin.h
@@ -8,7 +8,7 @@
 #define VG_NORIGINS 3
 
 __BEGIN_DECLS
-void		VG_Origin(struct vg *, int, double, double);
+void		VG_Origin(struct vg *, int, float, float);
 void		VG_OriginColor(struct vg *, int, int, int, int);
 void		VG_OriginRadius(struct vg *, int, float);
 __inline__ void VG_DrawOrigin(struct vg *);
